trace: move identical code to a common header
Contents of platform/trace/trace.h on all cAVS platforms are
identical, replace them with a common header.
