+++ src/platform/tigerlake/include/platform/trace/trace.h
@@ -10,27 +10,7 @@
 #ifndef __PLATFORM_TRACE_TRACE_H__
 #define __PLATFORM_TRACE_TRACE_H__
 
-#include <sof/lib/cpu.h>
-#include <sof/lib/mailbox.h>
-#include <sof/lib/memory.h>
-#include <stdint.h>
-
-#define PLATFORM_TRACEP_SECONDARY_CORE(x) \
-	(SRAM_REG_FW_TRACEP_SECONDARY_CORE_BASE + ((x) - 1) * 0x4)
-
-/* Platform defined trace code */
-static inline void platform_trace_point(uint32_t x)
-{
-	int cpu = cpu_get_id();
-	uint32_t offset;
-
-	if (cpu == PLATFORM_PRIMARY_CORE_ID)
-		offset = SRAM_REG_FW_TRACEP;
-	else
-		offset = PLATFORM_TRACEP_SECONDARY_CORE(cpu);
-
-	mailbox_sw_reg_write(offset, x);
-}
+#include <cavs/trace/trace.h>
 
 #endif /* __PLATFORM_TRACE_TRACE_H__ */
